Fix type conversion errors in Swift code for Windows builds
- Fix UInt conversion for compileFlags (swift_zlibCompileFlags returns uLong)
- Fix Int to z_off_t conversion for adler32Combine and crc32Combine functions
- These fixes resolve the Swift compilation errors on Windows
- Build now succeeds locally with these changes

+++ Sources/SwiftZlib/API/ZLib.swift
@@ -81,7 +81,7 @@ public enum ZLib {
 
     /// Get ZLib compile flags
     public static var compileFlags: UInt {
-        swift_zlibCompileFlags()
+        UInt(swift_zlibCompileFlags())
     }
 
     /// Get detailed ZLib compile flags information
